extract CurrencyIdConvert
